moved all structs to types.go
